qvolution-final-main/requirements.txt
numpy
pandas
torch
scikit-learn
scipy
matplotlib
openpyxl
perceval-quandela==1.1.0
merlinquantum==0.3.1
fastapi
uvicorn